+++ Changelog.md
@@ -36,6 +36,7 @@
 - API: Refactored IPasscodeConvertible to account for the new Wire Cutters functionality
 - API: New method IOwnable#ignoresOwner to check whether the owner should be ignored by this block entity
 - API: IOwnable#isOwnedBy now takes any Entity instead of only players
+- API: LinkableBlockEntity#createLinkedBlockAction has been renamed to LinkableBlockEntity#propagate
 - Fix: Duplication exploit involving keycards
 - Fix: Duplication exploit involving Reinforced Pistons
 - Fix: The name of certain SecurityCraft blocks is displayed within brackets by WTHIT and similar mods